Funciones de firebase movidas a index.js
