+++ src/lib/index.js
@@ -1,6 +1,17 @@
-// aqui exportaras las funciones que necesites
+import { initializeApp } from 'firebase/app';
+import { getAuth, signInWithEmailAndPassword, signInWithPopup, GoogleAuthProvider } from 'firebase/auth';
 
-export const myFunction = () => {
-  // aqui tu codigo
-  console.log('Hola mundo!');
+const firebaseConfig = {
+  apiKey: 'AIzaSyAF2gJx3-nMb9WaoJd82xBfmmtU0neZ2UA',
+  authDomain: 'nanaisocialnetwork.firebaseapp.com',
+  databaseURL: 'https://nanaisocialnetwork-default-rtdb.firebaseio.com',
+  projectId: 'nanaisocialnetwork',
+  storageBucket: 'nanaisocialnetwork.appspot.com',
+  messagingSenderId: '117090233074',
+  appId: '1:117090233074:web:1e82a71a00e02fdce8be6b',
 };
+
+const app = initializeApp(firebaseConfig);
+const auth = getAuth();
+
+export const crearUsuario = (email, contraseña) => signInWithEmailAndPassword(auth, email,contraseña);
